+++ aios/llm_core/adapter.py
@@ -1,98 +1,193 @@
+from aios.utils.id_generator import generator_tool_call_id
+from cerebrum.llm.communication import Response
+from litellm import completion
+import json
 
 from typing import Dict, Optional
-from aios.llm_core.cores.base import BaseLLM
-from aios.llm_core.registry import BACKEND_REGISTRY, MODEL_PREFIX_MAP
-from aios.llm_core.cores.local.hf import HfNativeLLM
+import time
+import re
 
 class LLMAdapter:
-    """Parameters for LLMs
+    """
+    The LLMAdapter class is an abstraction layer that represents the LLM
+    router. This router allows load-balancing of multiple varying endpoints so
+    that multiple requests can be handled at once.
 
     Args:
-        llm_name (str): Name of the LLMs
-        max_gpu_memory (dict, optional): Maximum GPU resources that can be allocated to the LLM. Defaults to None.
-        eval_device (str, optional): Evaluation device of binding LLM to designated devices for inference. Defaults to None.
-        max_new_tokens (int, optional): Maximum token length generated by the LLM. Defaults to 256.
-        log_mode (str, optional): Mode of logging the LLM processing status. Defaults to "console".
-        llm_backend (str, optional): Backend to use for speeding up open-source LLMs. Defaults to None. Choices are ["vllm", "ollama"]
+        llm_name (str or List[str])     : Name of the LLMs. If a string is
+                                          provided, then only one LLM will be
+                                          activated.
+        max_gpu_memory (dict, optional) : Maximum GPU resources that can be 
+                                          allocated to the LLMs.
+        eval_device (str, optional)     : Evaluation device of binding LLM to
+                                          designated devices for inference.
+        max_new_tokens (int, optional)  : Maximum token length generated by the
+                                          LLM. Defaults to 256.
+        log_mode (str, optional)        : Mode of logging the LLM processing
+                                          status. Defaults to "console".
+        llm_backend (str, optional)     : Backend to use for speeding up 
+                                          open-source LLMs. Defaults to None.
+                                          Choices are ["vllm", "ollama"]
     """
 
     def __init__(
         self,
-        llm_name: str,
+        llm_name: str | list[str],
         max_gpu_memory: Optional[Dict] = None,
         eval_device: Optional[str] = None,
         max_new_tokens: int = 256,
         log_mode: str = "console",
         llm_backend: Optional[str] = None,
         use_context_manager: bool = False,
-        api_key: str = None
+        api_key: str | list[str] | None = None
     ):
         """Initialize the LLM with the specified configuration.
         
         Args:
-            llm_name: Name of the LLM model to use
-            max_gpu_memory: Maximum GPU memory allocation per device
-            eval_device: Device to use for evaluation
-            max_new_tokens: Maximum number of new tokens to generate
-            log_mode: Logging mode ("console" or other options)
-            use_backend: Specific backend to use (if None, inferred from model name)
-            use_context_manager: Whether to use context manager
-            api_key: API Key for the LLM
+            llm_name            : Name of the LLM model to use
+            max_gpu_memory      : Maximum GPU memory allocation per device
+            eval_device         : Device to use for evaluation
+            max_new_tokens      : Maximum number of new tokens to generate
+            log_mode            : Logging mode ("console" or other options)
+            use_backend         : Specific backend to use (if None, inferred
+                                  from model name)
+            use_context_manager : Whether to use context manager
+            api_key             : DEPRECATED. This was originally used to store
+                                  an API Key for the LLM, but LiteLLM uses keys
+                                  directly from the process environment
+                                  variables making this needless.
         """
-        self.model: Optional[BaseLLM] = None
-        
-        # Common model parameters
-        model_params = {
-            'llm_name': llm_name,
-            'log_mode': log_mode,
-            'use_context_manager': use_context_manager,
-            'api_key': api_key
-        }
-        
-        # print(llm_name)
-        # print(llm_backend)
-        
-        # print(llm_name)
-        
-        # If backend is explicitly specified, use it
-        # print(f"llm backend: {llm_backend}")
-        if llm_backend and llm_backend in BACKEND_REGISTRY:
-            if llm_backend in ["vllm", "ollama", "huggingface"]:
-                model_params.pop('api_key')
-            model_class = BACKEND_REGISTRY[llm_backend]
-            self.model = model_class(**model_params)
-            # print(self.model)
-            return
+        self.llm_name            = llm_name if isinstance(llm_name, list) else [llm_name]
+        self.max_gpu_memory      = max_gpu_memory
+        self.eval_device         = eval_device
+        self.max_new_tokens      = max_new_tokens
+        self.log_mode            = log_mode
+        self.llm_backend         = llm_backend
+        self.use_context_manager = use_context_manager
 
-        # Try to infer backend from model name prefix
-        model_prefix = next(
-            (prefix for prefix in MODEL_PREFIX_MAP.keys() 
-            if llm_name.lower().startswith(prefix)),
-            None
-        )
-        
-        # print(model_prefix)
-        
-        if model_prefix:
-            inferred_backend = MODEL_PREFIX_MAP[model_prefix]
-            model_class = BACKEND_REGISTRY[inferred_backend]
-            self.model = model_class(**model_params)
-            return
-        
-        # print("Here")
-        # Default to HuggingFace native implementation if no specific backend is found
-        self.model = HfNativeLLM(**model_params)
-
-    def address_syscall(self,
-                        llm_syscall,
-                        temperature=0.0) -> None:
-        """Address request sent from the agent
+    def tool_calling_input_format(self, messages: list, tools: list) -> list:
+        """Integrate tool information into the messages for open-sourced LLMs
 
         Args:
-            agent_request: AgentProcess object that contains request sent from the agent
-            temperature (float, optional): Parameter to control the randomness of LLM output. Defaults to 0.0.
+            messages (list): messages with different roles
+            tools (list): tool information
         """
-        return self.model.address_syscall(llm_syscall, temperature)
+        prefix_prompt = (
+            "In and only in current step, you need to call tools. Available tools are: "
+        )
+        tool_prompt = json.dumps(tools)
+        suffix_prompt = "".join(
+            [
+                "Must call functions that are available. To call a function, respond "
+                "immediately and only with a list of JSON object of the following format:"
+                '{[{"name":"function_name_value","parameters":{"parameter_name1":"parameter_value1",'
+                '"parameter_name2":"parameter_value2"}}]}'
+            ]
+        )
 
+        # translate tool call message for models don't support tool call
+        for message in messages:
+            if "tool_calls" in message:
+                message["content"] = json.dumps(message.pop("tool_calls"))
+            elif message["role"] == "tool":
+                message["role"] = "user"
+                tool_call_id = message.pop("tool_call_id")
+                content = message.pop("content")
+                message["content"] = (
+                    f"The result of the execution of function(id :{tool_call_id}) is: {content}. "
+                )
 
-    
+        messages[-1]["content"] += prefix_prompt + tool_prompt + suffix_prompt
+        return messages
+
+    def parse_json_format(self, message: str) -> str:
+        json_array_pattern = r"\[\s*\{.*?\}\s*\]"
+        json_object_pattern = r"\{\s*.*?\s*\}"
+
+        match_array = re.search(json_array_pattern, message)
+        
+        # print(f"match_array: {match_array}")
+
+        if match_array:
+            json_array_substring = match_array.group(0)
+
+            try:
+                json_array_data = json.loads(json_array_substring)
+                return json.dumps(json_array_data)
+            except json.JSONDecodeError:
+                pass
+
+        match_object = re.search(json_object_pattern, message)
+
+        if match_object:
+            json_object_substring = match_object.group(0)
+
+            try:
+                json_object_data = json.loads(json_object_substring)
+                return json.dumps(json_object_data)
+            except json.JSONDecodeError:
+                pass
+        return "[]"
+
+    def parse_tool_calls(self, message):
+        # add tool call id and type for models don't support tool call
+        tool_calls = json.loads(self.parse_json_format(message))
+        for tool_call in tool_calls:
+            tool_call["id"] = generator_tool_call_id()
+            tool_call["type"] = "function"
+        return tool_calls
+
+    def address_syscall(
+        self,
+        llm_syscall,
+        temperature=0.0
+    ):
+        """
+        Address request sent from the agent
+
+        Args:
+            llm_syscall (LLMSyscall)      : LLMSyscall object that contains
+                                            request sent from the agent
+            temperature (float, optional) : Parameter to control the randomness
+                                            of LLM output. Defaults to 0.0.
+        """
+        messages = llm_syscall.query.messages
+        tools    = llm_syscall.query.tools
+        ret_type = llm_syscall.query.message_return_type
+
+        llm_syscall.set_status("executing")
+        llm_syscall.set_start_time(time.time())
+
+        #self.logger.log(
+        #    f"{llm_syscall.agent_name} is switched to executing.\n", level="executing"
+        #)
+
+        if self.use_context_manager:
+            return Response(
+                response_message="Context manager not supported yet",
+                finished=True,
+            )
+        else:
+            if tools:
+                messages = self.tool_calling_input_format(messages, tools)
+            
+            # TODO: Load balancing
+            res = str(completion(
+                model=self.llm_name[0],
+                messages=messages,
+                temperature=temperature,
+            ))
+
+            if tools:
+                tool_calls = self.parse_tool_calls(res)
+                if tool_calls:
+                    return Response(
+                        response_message=None,
+                        tool_calls=tool_calls,
+                        finished=True,
+                    )
+
+            if ret_type == "json":
+                res = self.parse_json_format(res)
+
+            return Response(response_message=res, finished=True)
